[XLA] Add a rounding_mode field to convert HLO to control float->int conversion.
Currently the kind of rounding to perform is unspecified and uncontrollable.
Adding a field to allow control over this that gives the user power to control
the kind of rounding to perform.

Available options are:
ROUND_TO_ZERO -> Round towards zero.
ROUND_NEAREST_NE -> Round to nearest ties to even
ROUND_NEAREST_AFZ -> Round to nearest ties away from zero

PiperOrigin-RevId: 413322977